+++ src/main/kotlin/com/young_zy/forum/controller/request/GiveSectionAdminRequest.kt
@@ -1,6 +1,6 @@
 package com.young_zy.forum.controller.request
 
 data class GiveSectionAdminRequest(
-        val sectionIds: List<Int>,
+        val sectionIds: List<Long>,
         val userIds: List<Long>
 )
